Split compound search into different files
