+++ src/app/_models/index.ts
@@ -3,5 +3,6 @@ export * from './user';
 export * from './loader-state';
 export * from './login-state';
 export * from './route-def';
+export * from './search-result';
 export * from './vendor-data/index';
 export * from './WDQSData';
