+++ src/app/_services/index.ts
@@ -1,3 +1,5 @@
 export * from './google-analytics.service';
 export * from './loader-state.service';
 export * from './login-state.service';
+export * from './search-results.service';
+export * from './WDQ.service';
