+++ src/app/compound-search/index.ts
@@ -1,0 +1,5 @@
+export * from './compound-search.component';
+export * from './compound-search-options/compound-search-options.component';
+export * from './search-box/search-box.component';
+export * from './search-result/search-result.component';
+export * from './search-results-table/search-results-table.component';
