+++ src/app/compound-search/search-box/search-box.component.html
@@ -1,0 +1,1 @@
+<input type="text" class="form-control" placeholder="Search: <Enter Drug Name or InChI key>" autofocus [(ngModel)]="searchQuery">
