+++ src/app/compound-search/search-result/search-result.component.html
@@ -1,0 +1,1 @@
+<app-search-results-table [result]="result"></app-search-results-table>
